refactor reminder and data evaluation service
