+++ src/app/shared/model/flowTypes.ts
@@ -21,7 +21,6 @@ export namespace FlowTypes {
     | "care_package_list"
     | "tour"
     | "habit_ideas"
-    | "reminder_list"
     | "template"
     | "component_defaults"
     // global data provides data to other modules, without namespacing (all top-level)
@@ -92,10 +91,6 @@ export namespace FlowTypes {
     flow_type: "care_package_list";
     rows: CarePackage[];
   }
-  export interface Reminder_list extends FlowTypeWithData {
-    flow_type: "reminder_list";
-    rows: Reminder_listRow[];
-  }
   export interface Data_list extends FlowTypeWithData {
     flow_type: "data_list";
     rows: Data_listRow[];
@@ -138,10 +133,7 @@ export namespace FlowTypes {
     rows?: Module_pageRow[];
   }
   /** all data_list type must provide a unique id for each row to allow */
-  interface Data_listRow {
-    id: string;
-    [key: string]: any;
-  }
+  export type Data_listRow<T = any> = { id: string } & T;
   export interface Habit_listRow extends Data_listRow {
     title: string;
     description: string;
@@ -166,7 +158,7 @@ export namespace FlowTypes {
     id: string;
     start_action?: Start_action;
     /** when tasks launch flows specify the type and name of flow. Only specific types are currently handled, as listed here */
-    flow_type?: "conversation" | "tips";
+    flow_type?: "template" | "conversation" | "tips";
     /** when tasks launch flows specify the type and name of flow */
     flow_name?: string;
     /** when tasks require additional paremeters, such as the name of a reward, provide here */
@@ -252,26 +244,29 @@ export namespace FlowTypes {
     habit_list: string[];
   }
 
-  export interface Reminder_listRow {
-    reminder_id: string;
-    /** start actions will be triggered in the task service so action types must match */
-    start_action: Start_action;
-    flow_type: Task_listRow["flow_type"];
-    /** args should match the name of the flow when being used with a start_new_flow task action */
-    start_action_args: string;
-    priority: number;
-    activation_condition_list: Reminder_conditionList[];
-    deactivation_condition_list: Reminder_conditionList[];
-    campaign_list: Reminder_campaign[];
+  export interface Campaign_listRow {
+    _id: string;
+    activation_condition_list: DataEvaluationCondition[];
+    deactivation_condition_list: DataEvaluationCondition[];
+    campaign_list: string[]; // ids of campaigns where to run
+    priority?: number; // higher numbers will be given more priority
+
+    // additional fields for current data_list but not required
+    click_action_list?: TemplateRowAction[];
+    icon?: string;
+    text?: string;
+
+    // placeholder for any extra fields to be added
+    [field: string]: any;
   }
-  export interface Reminder_conditionList {
+  export interface DataEvaluationCondition {
     /** specific defined actions that have individual methods to determine completion */
     condition_type: "field_evaluation" | "db_lookup";
     /** Condition args change depending on type, hard to enforce typing switch so just include type mapping */
     condition_args: {
       db_lookup?: {
         table_id: IDBTable;
-        filter: { field: string; value: string | number };
+        filter: { field: string; value: string | number | boolean };
         order?: "asc" | "desc";
         evaluate?: {
           operator: ">" | "<=";
@@ -290,7 +285,6 @@ export namespace FlowTypes {
     _cleaned?: string;
     _parsed?: string[][];
   }
-  type Reminder_campaign = "campaign_main" | "campaign_evening" | "campaign_morning";
 
   export interface Habit_ideas extends FlowTypeWithData {
     flow_type: "habit_ideas";
@@ -418,7 +412,9 @@ export namespace FlowTypes {
   export interface TemplateRowDynamicEvaluator {
     fullExpression: string;
     matchedExpression: string;
-    type: "local" | "field" | "fields" | "global" | "data";
+    // TODO CC 2021-05-15 - 'campaign' should be handled as a special case of data in the parser
+    // i.e. @data.campaign_list | evaluate_conditions | first (or similar)
+    type: "local" | "field" | "fields" | "global" | "data" | "campaign";
     fieldName: string;
   }
 
@@ -451,7 +447,8 @@ export namespace FlowTypes {
       | "style"
       | "close_pop_up"
       | "set_theme"
-      | "start_tour";
+      | "start_tour"
+      | "trigger_actions";
     args: string[];
     /** field populated for tracking the component that triggered the action */
     _triggeredBy?: TemplateRow;
